move loadLanguage to init.ts
